+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   `@lumx/vue`:
     -   Create the `Dialog` component
+    -   Create the `InfiniteScroll` utility component (available from `@lumx/vue/utils`)
 
 ### Changed